Add PInvoke/SizeParamIndex tests (dotnet/coreclr#19348)
* Add PInvoke/SizeParamIndex tests

* Fix compile warnings

* First pass cleaning up native code.

* Get ParamSizeIndex tests working xplat (excluding the BSTR array tests)

* Fix incorrect definition of ULONG off-windows.

* Disable reverse-PInvoke throwing test off-Windows.


Commit migrated from https://github.com/dotnet/coreclr/commit/243116831430aa73338b05cdeba94bbd347df0fe

+++ src/coreclr/tests/src/Common/Platform/platformdefines.h
@@ -39,7 +39,7 @@ typedef const WCHAR *LPCWSTR, *PCWSTR;
 typedef int HRESULT;
 #define LONGLONG long long
 #define ULONGLONG unsigned LONGLONG
-typedef unsigned long ULONG, *PULONG;
+typedef unsigned int ULONG, *PULONG;
 #define S_OK                    0x0
 #define SUCCEEDED(_hr)          ((HRESULT)(_hr) >= 0)
 #define FAILED(_hr)             ((HRESULT)(_hr) < 0)
